refactor vm2 txn factory
