+++ cmd/tx-factory/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod txn_generator;
+pub mod vm2_txn_lib;
